requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
phidata==2.3.30
googlesearch-python==1.2.3
pycountry==23.12.11
openai==1.12.0
duckduckgo-search==4.4.3
newspaper4k==0.9.3.1
lxml_html_clean==0.4.1
google-search-results==2.4.2
tavily-python==0.3.1
PyGithub==2.2.0
duckdb==0.10.0 